# api_spec.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "RASO Platform API",
                 "description":  "Research paper Automated Simulation \u0026 Orchestration Platform",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/jobs":  {
                                    "get":  {
                                                "summary":  "List Jobs",
                                                "description":  "List all jobs.",
                                                "operationId":  "list_jobs_api_jobs_get",
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "Successful Response",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {

                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          }
                                                              }
                                            },
                                    "post":  {
                                                 "summary":  "Submit Job",
                                                 "description":  "Submit a new video generation job.",
                                                 "operationId":  "submit_job_api_jobs_post",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/JobRequest"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/JobResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               }
                                             }
                                },
                  "/api/jobs/{job_id}":  {
                                             "get":  {
                                                         "summary":  "Get Job Status",
                                                         "description":  "Get job status.",
                                                         "operationId":  "get_job_status_api_jobs__job_id__get",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "job_id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Job Id"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/JobStatus"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     }
                                         },
                  "/api/jobs/{job_id}/download":  {
                                                      "get":  {
                                                                  "summary":  "Download Video",
                                                                  "description":  "Download generated video.",
                                                                  "operationId":  "download_video_api_jobs__job_id__download_get",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "job_id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "Job Id"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {

                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              }
                                                  }
              },
    "components":  {
                       "schemas":  {
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "JobRequest":  {
                                                          "properties":  {
                                                                             "paper_input":  {
                                                                                                 "$ref":  "#/components/schemas/PaperInput"
                                                                                             },
                                                                             "options":  {
                                                                                             "anyOf":  [
                                                                                                           {
                                                                                                               "additionalProperties":  true,
                                                                                                               "type":  "object"
                                                                                                           },
                                                                                                           {
                                                                                                               "type":  "null"
                                                                                                           }
                                                                                                       ],
                                                                                             "title":  "Options"
                                                                                         }
                                                                         },
                                                          "type":  "object",
                                                          "required":  [
                                                                           "paper_input"
                                                                       ],
                                                          "title":  "JobRequest",
                                                          "description":  "Job submission request."
                                                      },
                                       "JobResponse":  {
                                                           "properties":  {
                                                                              "job_id":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Job Id"
                                                                                         },
                                                                              "status":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Status"
                                                                                         },
                                                                              "created_at":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time",
                                                                                                 "title":  "Created At"
                                                                                             }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "job_id",
                                                                            "status",
                                                                            "created_at"
                                                                        ],
                                                           "title":  "JobResponse",
                                                           "description":  "Job submission response."
                                                       },
...
                                       "PaperInput":  {
                                                          "properties":  {
                                                                             "type":  {
                                                                                          "$ref":  "#/components/schemas/PaperInputType",
                                                                                          "description":  "Type of paper input"
                                                                                      },
                                                                             "content":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Content",
                                                                                             "description":  "Paper content (title, URL, or base64 PDF)"
                                                                                         },
                                                                             "options":  {
                                                                                             "anyOf":  [
                                                                                                           {
                                                                                                               "additionalProperties":  true,
                                                                                                               "type":  "object"
                                                                                                           },
                                                                                                           {
                                                                                                               "type":  "null"
                                                                                                           }
                                                                                                       ],
                                                                                             "title":  "Options",
                                                                                             "description":  "Processing options"
                                                                                         }
                                                                         },
                                                          "type":  "object",
                                                          "required":  [
                                                                           "type",
                                                                           "content"
                                                                       ],
                                                          "title":  "PaperInput",
                                                          "description":  "Input specification for a research paper."
                                                      },
                                       "PaperInputType":  {
                                                              "type":  "string",
                                                              "enum":  [
                                                                           "title",
                                                                           "arxiv",
                                                                           "pdf"
                                                                       ],
                                                              "title":  "PaperInputType",
                                                              "description":  "Types of paper input supported by RASO."
                                                          },
...
                                   }
                   }
}